wip: adding history backend sqlite
