+++ xonsh/built_ins.py
@@ -23,7 +23,7 @@ import collections.abc as cabc
 
 from xonsh.ast import AST
 from xonsh.lazyasd import LazyObject, lazyobject
-from xonsh.history import History
+from xonsh.history import get_history_backend
 from xonsh.inspectors import Inspector
 from xonsh.aliases import Aliases, make_default_aliases
 from xonsh.environ import Env, default_env, locate_binary
@@ -1151,14 +1151,15 @@ def load_builtins(execer=None, config=None, login=False, ctx=None):
     builtins.events = events
 
     # sneak the path search functions into the aliases
-    # Need this inline/lazy import here since we use locate_binary that relies on __xonsh_env__ in default aliases
+    # Need this inline/lazy import here since we use locate_binary that
+    # relies on __xonsh_env__ in default aliases
     builtins.default_aliases = builtins.aliases = Aliases(make_default_aliases())
     if login:
         builtins.aliases.update(load_foreign_aliases(issue_warning=False))
     # history needs to be started after env and aliases
     # would be nice to actually include non-detyped versions.
-    builtins.__xonsh_history__ = History(env=env.detype(),
-                                         ts=[time.time(), None], locked=True)
+    builtins.__xonsh_history__ = get_history_backend(
+        env=env.detype(), ts=[time.time(), None], locked=True)
     atexit.register(_lastflush)
     for sig in AT_EXIT_SIGNALS:
         resetting_signal_handle(sig, _lastflush)
