+++ xonsh/replay.py
@@ -7,7 +7,7 @@ import collections.abc as cabc
 from xonsh.tools import swap
 from xonsh.lazyjson import LazyJSON
 from xonsh.environ import Env
-from xonsh.history import History
+from xonsh.history import get_history_backend
 from xonsh.history import _hist_info
 
 
@@ -50,8 +50,9 @@ class Replayer(object):
         shell = builtins.__xonsh_shell__
         re_env = self._lj['env'].load()
         new_env = self._merge_envs(merge_envs, re_env)
-        new_hist = History(env=new_env.detype(), locked=True, ts=[time.time(), None],
-                           gc=False, filename=target)
+        new_hist = get_history_backend(
+            env=new_env.detype(), locked=True, ts=[time.time(), None],
+            gc=False, filename=target)
         with swap(builtins, '__xonsh_env__', new_env), swap(builtins, '__xonsh_history__', new_hist):
             for cmd in self._lj['cmds']:
                 inp = cmd['inp']
